[v4l2_bridge] Add to set fps
This patch adds to set fps. This is implemented using usleep()
instead of using v4l2 framerate ioctl to be able to work even
on devices not supporting fps configuration. How often to dequeue
from input pipeline is decided based on fps, and if fps = -1,
the bridge application runs on free run mode (queue/dequeue a buffer,
whenever it's ready)